archive/gbif_species_trainer/requirements.txt
pygbif
scikit-learn
webdataset
wandb
torchsummary
tensorflow
matplotlib
keras
numpy
tqdm
pandas
timm
opencv-python
seaborn
python-dwca-reader
python-dotenv
SQLAlchemy
psycopg2-binary
black